Fix Microsoft Defender for Cloud Apps breaking login
If Microsoft Defender for Cloud Apps is enabled in Windows the requests will be proxied
through mcas.ms and therefore we never catch the SAMLResponse. On top of that even
if we would catch the SAMLResponse from the request body it will be encrypted and useless.
Instead we get the SAMLResponse from the from input in role selection web page.